+++ XsdToJavaAPIRoot/FasterApi/XsdAsmFaster/src/main/java/org/xmlet/xsdasmfaster/classes/infrastructure/RestrictionValidator.java
@@ -2,6 +2,9 @@ package org.xmlet.xsdasmfaster.classes.infrastructure;
 
 import java.util.List;
 
+/**
+ * THIS VALIDATOR DOESN'T VALIDATE THE WHITESPACE RESTRICTION.
+ */
 public final class RestrictionValidator {
 
     private RestrictionValidator() {}
@@ -30,13 +33,13 @@ public final class RestrictionValidator {
         }
     }
 
-    public static void validateMaxExclusive(int maxExclusive, double value){
+    public static void validateMaxExclusive(double maxExclusive, double value){
         if (value >= maxExclusive){
             throw new RestrictionViolationException("Violation of maxExclusive restriction, value should be lesser than " + maxExclusive);
         }
     }
 
-    public static void validateMaxInclusive(int maxInclusive, double value){
+    public static void validateMaxInclusive(double maxInclusive, double value){
         if (value > maxInclusive){
             throw new RestrictionViolationException("Violation of maxInclusive restriction, value should be lesser or equal to " + maxInclusive);
         }
@@ -54,13 +57,13 @@ public final class RestrictionValidator {
         }
     }
 
-    public static void validateMinExclusive(int minExclusive, double value){
+    public static void validateMinExclusive(double minExclusive, double value){
         if (value <= minExclusive){
             throw new RestrictionViolationException("Violation of minExclusive restriction, value should be greater than " + minExclusive);
         }
     }
 
-    public static void validateMinInclusive(int minInclusive, double value){
+    public static void validateMinInclusive(double minInclusive, double value){
         if (value < minInclusive){
             throw new RestrictionViolationException("Violation of minInclusive restriction, value should be greater or equal to " + minInclusive);
         }
